Fix some issues and move class definitions to different places
