+++ include/NullGameEngine/Script.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <NullGameEngine.hpp>
+#include <Component.hpp>
+
+namespace null {
+    class Script : public Component{
+    public:
+        virtual void start();
+        virtual void update();
+    };
+}
